Correcting response for location

+++ API/Controllers/LocationController.cs
@@ -27,12 +27,19 @@ namespace API.Controllers
 
 
         [HttpGet("get/location")]
-        public async Task<IEnumerable<LocationDto>> GetLocation(Guid? id)
+        public async Task<Response<IEnumerable<LocationDto>>> GetLocation(Guid? id)
         {
             var dp_params = new DynamicParameters();
+            Response<IEnumerable<LocationDto>> response = new Response<IEnumerable<LocationDto>>();
             dp_params.Add("@Id ", id, DbType.Guid);
-            return await _db.GetAll<LocationDto>("GetLocation", dp_params);
+            var newData = await _db.GetAll<LocationDto>("GetLocation", dp_params); ;
 
+            if (newData != null)
+            {
+                response.Code = 200;
+                response.Data = newData;
+            }
+            return response;
         }
 
         [HttpPost("create/location")]
@@ -45,32 +52,44 @@ namespace API.Controllers
             {
                 string sqlCommand = "insert into Location(LocationName)" +
                     "values (@LocationName)";
-                
-                if(await db.ExecuteAsync(sqlCommand, dp_params, null, null, CommandType.Text) == 1)
+
+                if (await db.ExecuteAsync(sqlCommand, dp_params, null, null, CommandType.Text) == 1)
                 {
                     sqlCommand = "select top 1 * from Location order by Id desc";
-                    var newData = await db.QueryFirstAsync<LocationDto>(sqlCommand, null, null, null, CommandType.Text);
+                    var newData = await db.QueryFirstAsync<LocationDto>(sqlCommand, dp_params, null, null, CommandType.Text);
                     if (newData != null)
                     {
                         response.Code = 200;
                         response.Data = newData;
-                    }    
-                }   
+                    }
+                }
             };
             return response;
         }
 
         [HttpDelete("delete/location")]
-        public async Task<ActionResult> DeleteLocaion(Guid id)
+        public async Task<Response<LocationDto>> DeleteLocation(Guid id)
         {
             var dp_params = new DynamicParameters();
+            Response<LocationDto> response = new Response<LocationDto>();
             dp_params.Add("@Id", id, DbType.Guid);
+
             using (IDbConnection db = new SqlConnection(_config.GetConnectionString("DefaultConnection")))
             {
-                string sqlCommand = "delete from Location where Id = @Id";
-                await db.ExecuteAsync(sqlCommand, dp_params, null, null, CommandType.Text);
+                string sqlCommand = "select * from Location where Id = @Id";
+                var oldData = await db.QueryFirstAsync<LocationDto>(sqlCommand, dp_params, null, null, CommandType.Text);
+
+                if (oldData != null)
+                {
+                    sqlCommand = "delete from Location where Id = @Id";
+                    await db.ExecuteAsync(sqlCommand, dp_params, null, null, CommandType.Text);
+
+                    response.Code = 200;
+                    response.Data = oldData;
+                }
             };
-            return Ok("Success");
+
+            return response;
         }
     }
 }
